Refactor : 인터페이스 정의 부분들은 따로 type 폴더에 정의
