+++ src/types/user.ts
@@ -1,0 +1,6 @@
+export interface User {
+    id: number;
+    name: string;
+    honorCount: number;
+    profileImage: string;
+}
